+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [1.4.2] - 2026-04-10
+
+### Fixed
+
+- Running `REINDEX CONCURRENTLY` (or `reindexdb --concurrently`) on a database
+  containing continuous views could crash all PipelineDB background workers with
+  "could not open relation with OID" errors, bringing down the entire server.
+  Subsequent stream inserts after the crash would also fail until a full restart.
+  Fixed by detecting when matrel index OIDs have been replaced and updating both
+  the internal catalog cache and the combiner's cached query plans accordingly
+  (TTDB-1505).
+- Incompatibility with the final Postgres 18 release leading to a compilation error
+
 ## [1.4.1] - 2025-09-16
 
 ### Added
@@ -145,10 +160,14 @@ This is the first formal Tantor release.
 
 This was the last release from [pipelinedb.com](https://www.pipelinedb.com) ([GitHub](https://github.com/pipelinedb/pipelinedb)).
 
-[unreleased]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.3.4-REL_17...REL_17_STABLE
-[1.3.4]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.3.3-REL_16...1.3.4-REL_17
-[1.3.3]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.3.2-REL_16...1.3.3-REL_16
-[1.3.2]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.3.1-REL_16...1.3.2-REL_16
-[1.3.1]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.3.0-REL_16...1.3.1-REL_16
-[1.3.0]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.0.0-13...1.3.0-REL_16
+[unreleased]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.4.2-REL_18...REL_18_STABLE
+[1.4.2]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.4.1-REL_18...1.4.2-REL_18
+[1.4.1]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.4.0-REL_18...1.4.1-REL_18
+[1.4.0]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.3.5-REL_18...1.4.0-REL_18
+[1.3.5]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.3.4-REL_18...1.3.5-REL_18
+[1.3.4]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.3.3-REL_18...1.3.4-REL_18
+[1.3.3]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.3.2-REL_18...1.3.3-REL_18
+[1.3.2]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.3.1-REL_18...1.3.2-REL_18
+[1.3.1]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.3.0-REL_18...1.3.1-REL_18
+[1.3.0]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/compare/1.0.0-13...1.3.0-REL_18
 [1.0.0-13]: https://gitlab.tantorlabs.ru/database/pipelinedb/-/tags/1.0.0-13